+++ lib/pages/tagesabschluss_schritt2/ui/schritt2_gruppen_orchestrierung.dart
@@ -2,6 +2,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:kino_bar_app/pages/tagesabschluss_schritt2/sections/schritt2_anmerkung_section.dart';
 import 'package:kino_bar_app/pages/tagesabschluss_schritt2/sections/schritt2_differenz_anfangsbestand_section.dart';
+import 'package:kino_bar_app/pages/tagesabschluss_schritt2/sections/schritt2_ec_beleg_sub_kacheln.dart';
 import 'package:kino_bar_app/pages/tagesabschluss_schritt2/sections/schritt2_ec_belege_kachel_section.dart';
 import 'package:kino_bar_app/pages/tagesabschluss_schritt2/sections/schritt2_kino_soll_ausgaben_section.dart';
 import 'package:kino_bar_app/pages/tagesabschluss_schritt2/sections/schritt2_kopf_section.dart';
@@ -107,6 +108,54 @@ class Schritt2GruppenOrchestrierung {
       onBelegdatenBearbeitenTap: onBelegdatenBearbeitenTap,
       onEcBelegHinzufuegen: onEcBelegHinzufuegen,
       belegInhalt: belegInhalt,
+    );
+  }
+
+  Widget baueEcBelegSubKacheln({
+    required int belegAnzahl,
+    required List<int> belegIds,
+    required List<bool> aufgeklapptListe,
+    required void Function(int belegIndex) onToggleAufgeklappt,
+    required List<bool> editModusListe,
+    required int? scanBelegIndex,
+    required List<TextEditingController> labelController,
+    required List<FocusNode> labelFocusNode,
+    required List<String> labels,
+    required List<int> betraegeCent,
+    required bool Function(int belegIndex) tidUnleserlich,
+    required bool scanLaeuft,
+    required void Function(int belegIndex) onScanStarten,
+    required void Function(int belegIndex, String wert) onTidGeaendert,
+    required void Function(int belegIndex) onBestaetigtEntfernen,
+    required bool Function(int belegIndex) hatZahlungsartZeilen,
+    required bool Function(int belegIndex) hatScanStattgefunden,
+    required Widget Function(int belegIndex) baueZahlungsartenTabelle,
+    required Widget Function(int belegIndex) baueMetadatenBlock,
+    required void Function(int belegIndex) onManuellBearbeitenAktivieren,
+    required void Function(int belegIndex) onFertig,
+  }) {
+    return Schritt2EcBelegSubKacheln(
+      belegAnzahl: belegAnzahl,
+      belegIds: belegIds,
+      aufgeklapptListe: aufgeklapptListe,
+      onToggleAufgeklappt: onToggleAufgeklappt,
+      editModusListe: editModusListe,
+      scanBelegIndex: scanBelegIndex,
+      labelController: labelController,
+      labelFocusNode: labelFocusNode,
+      labels: labels,
+      betraegeCent: betraegeCent,
+      tidUnleserlich: tidUnleserlich,
+      scanLaeuft: scanLaeuft,
+      onScanStarten: onScanStarten,
+      onTidGeaendert: onTidGeaendert,
+      onBestaetigtEntfernen: onBestaetigtEntfernen,
+      hatZahlungsartZeilen: hatZahlungsartZeilen,
+      hatScanStattgefunden: hatScanStattgefunden,
+      baueZahlungsartenTabelle: baueZahlungsartenTabelle,
+      baueMetadatenBlock: baueMetadatenBlock,
+      onManuellBearbeitenAktivieren: onManuellBearbeitenAktivieren,
+      onFertig: onFertig,
     );
   }
 }